+++ ipc/shmem_ipc/common.c
@@ -4,8 +4,10 @@
 #include <fcntl.h>
 #include <pthread.h>
 #include <unistd.h>
-#include "LINF/sym_all.h"
+
 #include <time.h>
+
+
 
 /*
 Function that initalize the server and workspace
@@ -87,89 +89,13 @@ void* connect_server(){
 	return shared_memory;
 }
 
-void* job_buffer_thread(void *job_buffer){
-	
-	job_buffer_t *request_job_buffer = job_buffer;
-
-	//stub hardcoded variables
-	FILE * fp = NULL;
-	int fd = -1;
-	char * last_filename = "";
-
-	clock_t end = clock() + 1*CLOCKS_PER_SEC;
-
-	#ifdef ELEVATED
-	ksys_write_t ksys_write = (ksys_write_t)sym_get_fn_address("ksys_write");
-	getppid_t getppid_elevated = (getppid_t)sym_get_fn_address((char*)"__x64_sys_getppid");
-	sym_elevate();
-	#endif
-
-	printf("thread started at %p\n", job_buffer);
-		// ever running thread
-		//wait{ for request
-	LOOP:	
-		while (request_job_buffer->status != JOB_REQUESTED) {
-			if ((request_job_buffer->status == SERVER_KILL_SIGNAL) || (clock() > end)){
-				#ifdef ELEVATED
-				sym_lower();
-				#endif
-				pthread_exit (NULL);
-			}
-		}
-		int i = 0;
-		switch(request_job_buffer->cmd){
-
-			case CMD_WRITE: {
-				if ((fd == -1) | strcmp(request_job_buffer->filename, last_filename)){
-					if (fd != -1){
-						close(fd);
-					}
-					fp = fopen(request_job_buffer->filename, "a");
-					fd = fileno(fp);
-					last_filename = request_job_buffer->filename;
-				}
-				
-				#ifdef ELEVATED
-				i++;
-				if (i % 2000 == 0){
-					write(fd, request_job_buffer->buf, request_job_buffer->buf_len);
-				}else{
-					ksys_write(fd, request_job_buffer->buf, request_job_buffer->buf_len);
-				}
-				#else
-				write(fd, request_job_buffer->buf, request_job_buffer->buf_len);
-				#endif
-				request_job_buffer->response = 1;
-				break;
-			}
-
-			case CMD_GETPPID: {
-				#ifdef ELEVATED
-				getppid_elevated();
-				#else
-				getppid();
-				#endif
-				request_job_buffer->response = 1;
-				break;
-			}
-			default: {
-				break;
-			}
-
-		}
-		
-		request_job_buffer->status = JOB_COMPLETED;
-
-		goto LOOP;
-}
-
-int server_write(job_buffer_t * job_buffer, char * floc, char * buf, size_t len){
+int server_write(job_buffer_t * job_buffer, char * buf, size_t len){
 	//very stub write just doing the memory transfer
 
 	job_buffer->cmd = CMD_WRITE;
 	job_buffer->buf_len = len;
 	strcpy(job_buffer->buf, buf);
-	strcpy(job_buffer->filename, floc);
+	//strcpy(job_buffer->filename, floc);
 	//send request
 	job_buffer->status = JOB_REQUESTED;
 
@@ -185,4 +111,3 @@ int server_getppid(job_buffer_t * job_buffer){
 	while (job_buffer->status != JOB_COMPLETED);
 	return job_buffer->response;
 }
-
