+++ ipc/shmem_ipc/results.csv
@@ -1,21 +1,41 @@
 i, latency, apporach, elevate
-0, 0.132553, independent, unelevated
-0, 0.229324, server, elevated
-1, 0.131954, independent, unelevated
-1, 0.142480, server, elevated
-2, 0.131908, independent, unelevated
-2, 0.148874, server, elevated
-3, 0.131809, independent, unelevated
-3, 0.145065, server, elevated
-4, 0.132736, independent, unelevated
-4, 0.142809, server, elevated
-5, 0.130127, independent, unelevated
-5, 0.149237, server, elevated
-6, 0.132099, independent, unelevated
-6, 0.143554, server, elevated
-7, 0.133021, independent, unelevated
-7, 0.144180, server, elevated
-8, 0.131678, independent, unelevated
-8, 0.142734, server, elevated
-9, 0.131746, independent, unelevated
-9, 0.143256, server, elevated
+0, 0.192564, independent, unelevated
+0, 0.167321, independent, elevated
+0, 0.248828, server, unelevated
+0, 0.232244, server, elevated
+1, 0.206126, independent, unelevated
+1, 0.177140, independent, elevated
+1, 0.232736, server, unelevated
+1, 0.238237, server, elevated
+2, 0.208095, independent, unelevated
+2, 0.177120, independent, elevated
+2, 0.460589, server, unelevated
+2, 0.239008, server, elevated
+3, 0.207988, independent, unelevated
+3, 0.178809, independent, elevated
+3, 0.465073, server, unelevated
+3, 0.443988, server, elevated
+4, 0.206334, independent, unelevated
+4, 0.178221, independent, elevated
+4, 0.452739, server, unelevated
+4, 0.441562, server, elevated
+5, 0.207456, independent, unelevated
+5, 0.176394, independent, elevated
+5, 0.254249, server, unelevated
+5, 0.444638, server, elevated
+6, 0.206259, independent, unelevated
+6, 0.176778, independent, elevated
+6, 0.248193, server, unelevated
+6, 0.238648, server, elevated
+7, 0.206761, independent, unelevated
+7, 0.177209, independent, elevated
+7, 0.233877, server, unelevated
+7, 0.250575, server, elevated
+8, 0.205354, independent, unelevated
+8, 0.176286, independent, elevated
+8, 0.244943, server, unelevated
+8, 0.243510, server, elevated
+9, 0.206878, independent, unelevated
+9, 0.175906, independent, elevated
+9, 0.226565, server, unelevated
+9, 0.239683, server, elevated
